fix: relocate auth-failure payload helper to avoid unused export
