+++ src/agents/embedded-agent-runner/run/terminal-resolution.ts
@@ -22,6 +22,7 @@ import {
   hasAttemptTerminalState,
   hasYieldContinuationEvidence,
   resolveAttemptReplayMetadata,
+  resolveAuthFailurePayloadText,
   resolveEmptyResponseRetryInstruction,
   resolveIncompleteTurnPayloadText,
   resolveReasoningOnlyRetryInstruction,
@@ -32,20 +33,6 @@ import {
   shouldTreatEmptyAssistantReplyAsSilent,
   YIELD_DIAGNOSTIC_TEXT,
 } from "./incomplete-turn.js";
-
-export function resolveAuthFailurePayloadText(params: {
-  assistantProfileFailureReason?: AuthProfileFailureReason | null;
-  provider: string;
-  modelId: string;
-}): string | undefined {
-  if (
-    params.assistantProfileFailureReason !== "auth" &&
-    params.assistantProfileFailureReason !== "auth_permanent"
-  ) {
-    return undefined;
-  }
-  return `Authentication failed for model ${params.provider}/${params.modelId}. Please check your API key or switch to a different model.`;
-}
 import type { RunEmbeddedAgentParams } from "./params.js";
 import {
   MAX_BEFORE_AGENT_FINALIZE_REVISIONS,
@@ -296,7 +283,7 @@ export async function resolveEmbeddedRunTerminal(input: {
   }
   const incompleteTurnText = emptyAssistantReplyIsSilent
     ? null
-    : resolveAuthFailurePayloadText({
+    : (resolveAuthFailurePayloadText({
         assistantProfileFailureReason: input.assistantProfileFailureReason,
         provider: input.provider,
         modelId: input.modelId,
@@ -308,7 +295,7 @@ export async function resolveEmbeddedRunTerminal(input: {
         timedOut: input.terminalTimedOut,
         hadPotentialSideEffects: input.replayState.hadPotentialSideEffects,
         attempt,
-      });
+      }));
   const incompleteTurnFallbackSafe = Boolean(
     incompleteTurnText &&
     !input.terminalInterrupted &&
